pyproject.toml
[project]
name = "autosummary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.14.0",
    "llama-index-core>=0.13.6",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-llms-google-genai>=0.3.1",
    "llama-index-utils-workflow>=0.4.1",
    "llama-index-workflows>=1.3.0",
    "load-dotenv>=0.1.0",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]
